+++ Package.swift
@@ -1,0 +1,19 @@
+// swift-tools-version:5.1
+// The swift-tools-version declares the minimum version of Swift required to build this package.
+
+import PackageDescription
+
+let package = Package(
+    name: "Qonversion",
+    products: [
+        .library(
+            name: "Qonversion",
+            targets: ["Qonversion"]),
+    ],
+    targets: [
+        .target(
+            name: "Qonversion",
+            exclude: [],
+            publicHeadersPath: "."),
+    ]
+)
