+++ Qonversion.podspec
@@ -9,8 +9,8 @@ Pod::Spec.new do |s|
   s.homepage         = 'https://github.com/qonversion/qonversion-ios-sdk'
   s.license          = { :type => 'MIT', :file => 'LICENSE' }
   s.author           = { 'Qonversion Inc.' => 'hi@qonversion.io' }
+  s.source_files     = 'Sources/Qonversion/*.{h,m}''
   s.source           = { :git => 'https://github.com/qonversion/qonversion-ios-sdk.git', :tag => s.version.to_s }
   s.ios.deployment_target = '9.0'
-  s.source_files = 'Sources/**/*'
 
 end
