--- Sources/Network/Models/QNProduct+Protected.h
@@ -1,7 +1,0 @@
-#import "QNProduct.h"
-
-@interface QNProduct (Protected)
-
-@property (nonatomic, strong, readwrite, nullable) SKProduct *skProduct;
-
-@end
