+++ Sources/Qonversion/QNAPIClient.h
@@ -1,5 +1,4 @@
-#import "Qonversion.h"
-
+#import "Foundation/Foundation.h"
 @class SKProduct, SKPaymentTransaction;
 
 typedef void (^QNAPIClientCompletionHandler)(NSDictionary * _Nullable dict, NSError * _Nullable error);
@@ -24,7 +23,7 @@ NS_ASSUME_NONNULL_BEGIN
 
 - (void)properties:(NSDictionary *)properties completion:(QNAPIClientCompletionHandler)completion;
 
-- (void)attributionRequest:(QNAttributionProvider)provider
+- (void)attributionRequest:(NSInteger)provider
                       data:(NSDictionary *)data
                 completion:(QNAPIClientCompletionHandler)completion;
