+++ Sources/Qonversion/QNAPIClient.m
@@ -1,9 +1,9 @@
 #import "QNAPIClient.h"
+#import "QNConstants.h"
 #import "QNRequestBuilder.h"
 #import "QNRequestSerializer.h"
 #import "QNDevice.h"
 #import "QNMapper.h"
-#import "QNConstants.h"
 #import "QNErrors.h"
 #import "QNMapperObject.h"
 #import "QNUtils.h"
