+++ Sources/Qonversion/QNAttributionManager.h
@@ -1,11 +1,10 @@
 #import <Foundation/Foundation.h>
-#import "Qonversion.h"
 
 NS_ASSUME_NONNULL_BEGIN
 
 @interface QNAttributionManager : NSObject
 
-- (void)addAttributionData:(NSDictionary *)data fromProvider:(QNAttributionProvider)provider;
+- (void)addAttributionData:(NSDictionary *)data fromProvider:(NSInteger)provider;
 
 @end
