+++ Sources/Qonversion/QNConstants.m
