+++ Sources/Qonversion/QNDevice.h
