+++ Sources/Qonversion/QNDevice.m
@@ -1,6 +1,13 @@
 #import "QNDevice.h"
 #import "QNConstants.h"
+#if !TARGET_OS_OSX
 #import <UIKit/UIKit.h>
+#else
+#import <Cocoa/Cocoa.h>
+#import <net/if.h>
+#import <net/if_dl.h>
+#endif
+
 
 #import <sys/sysctl.h>
 #import <sys/types.h>
@@ -43,7 +50,15 @@
 
 - (NSString *)osVersion {
   if (!_osVersion) {
-    _osVersion = [[UIDevice currentDevice] systemVersion];
+    #if !TARGET_OS_OSX
+       _osVersion = [[UIDevice currentDevice] systemVersion];
+       #else
+       NSOperatingSystemVersion systemVersion = [[NSProcessInfo processInfo] operatingSystemVersion];
+       _osVersion = [NSString stringWithFormat:@"%ld.%ld.%ld",
+                     systemVersion.majorVersion,
+                     systemVersion.minorVersion,
+                     systemVersion.patchVersion];
+       #endif
   }
   return _osVersion;
 }
@@ -278,7 +293,12 @@
 }
 
 + (NSString*)getVendorID:(int) maxAttempts {
-  NSString *identifier = [[[UIDevice currentDevice] identifierForVendor] UUIDString];
+  #if !TARGET_OS_OSX
+      NSString *identifier = [[[UIDevice currentDevice] identifierForVendor] UUIDString];
+  #else
+      NSString *identifier = [self getMacAddress];
+  #endif
+  
   if (identifier == nil && maxAttempts > 0) {
     // Try again every 5 seconds
     [NSThread sleepForTimeInterval:5.0];
@@ -450,4 +470,74 @@
   if ([platform hasPrefix:@"Xserve"])             return @"Xserve";
   return platform;
 }
+
+#if TARGET_OS_OSX
++ (NSString *)getMacAddress {
+    int                 mgmtInfoBase[6];
+    char                *msgBuffer = NULL;
+    size_t              length;
+    unsigned char       macAddress[6];
+    struct if_msghdr    *interfaceMsgStruct;
+    struct sockaddr_dl  *socketStruct;
+    NSString            *errorFlag = NULL;
+    bool                msgBufferAllocated = false;
+
+    // Setup the management Information Base (mib)
+    mgmtInfoBase[0] = CTL_NET;        // Request network subsystem
+    mgmtInfoBase[1] = AF_ROUTE;       // Routing table info
+    mgmtInfoBase[2] = 0;
+    mgmtInfoBase[3] = AF_LINK;        // Request link layer information
+    mgmtInfoBase[4] = NET_RT_IFLIST;  // Request all configured interfaces
+
+    // With all configured interfaces requested, get handle index
+    if ((mgmtInfoBase[5] = if_nametoindex("en0")) == 0) {
+        errorFlag = @"if_nametoindex failure";
+    } else {
+        // Get the size of the data available (store in len)
+        if (sysctl(mgmtInfoBase, 6, NULL, &length, NULL, 0) < 0) {
+            errorFlag = @"sysctl mgmtInfoBase failure";
+        } else {
+            // Alloc memory based on above call
+            if ((msgBuffer = malloc(length)) == NULL) {
+                errorFlag = @"buffer allocation failure";
+            } else {
+                msgBufferAllocated = true;
+                // Get system information, store in buffer
+                if (sysctl(mgmtInfoBase, 6, msgBuffer, &length, NULL, 0) < 0) {
+                    errorFlag = @"sysctl msgBuffer failure";
+                }
+            }
+        }
+    }
+
+    // Before going any further...
+    if (errorFlag != NULL) {
+        AMPLITUDE_LOG(@"Cannot detect mac address. Error: %@", errorFlag);
+        if (msgBufferAllocated) {
+            free(msgBuffer);
+        }
+        return nil;
+    }
+
+    // Map msgbuffer to interface message structure
+    interfaceMsgStruct = (struct if_msghdr *) msgBuffer;
+
+    // Map to link-level socket structure
+    socketStruct = (struct sockaddr_dl *) (interfaceMsgStruct + 1);
+
+    // Copy link layer address data in socket structure to an array
+    memcpy(&macAddress, socketStruct->sdl_data + socketStruct->sdl_nlen, 6);
+
+    // Read from char array into a string object, into traditional Mac address format
+    NSString *macAddressString = [NSString stringWithFormat:@"%02X%02X%02X%02X%02X%02X",
+                                  macAddress[0], macAddress[1], macAddress[2],
+                                  macAddress[3], macAddress[4], macAddress[5]];
+
+    // Release the buffer memory
+    free(msgBuffer);
+
+    return macAddressString;
+}
+#endif
+
 @end
