+++ Sources/Qonversion/QNLaunchResult+Protected.h
@@ -1,3 +1,4 @@
+#import "Foundation/Foundation.h"
 #import "QNLaunchResult.h"
 
 @interface QNLaunchResult (Protected)
