+++ Sources/Qonversion/QNLaunchResult.h
@@ -1,0 +1,64 @@
+#import <Foundation/Foundation.h>
+
+@class QNPermission, QNProduct;
+
+typedef NS_ENUM(NSInteger, QNAttributionProvider) {
+  QNAttributionProviderAppsFlyer = 0,
+  QNAttributionProviderBranch,
+  QNAttributionProviderAdjust,
+  QNAttributionProviderApple
+} NS_SWIFT_NAME(Qonversion.AttributionProvider);
+
+/**
+ Qonversion Defined User Properties
+ We defined some common case properties and provided API for adding them
+ @see [Product Center](https://qonversion.io/docs/defined-user-properties)
+ */
+typedef NS_ENUM(NSInteger, QNProperty) {
+  QNPropertyEmail = 0,
+  QNPropertyName,
+  QNPropertyAppsFlyerUserID,
+  QNPropertyAdjustUserID,
+  QNPropertyKochavaDeviceID
+} NS_SWIFT_NAME(Qonversion.Property);
+
+
+NS_SWIFT_NAME(Qonversion.LaunchResult)
+@interface QNLaunchResult : NSObject <NSCoding>
+
+/**
+ Qonversion User Identifier
+ */
+@property (nonatomic, copy, readonly) NSString *uid;
+
+/**
+ Original Server response time
+ */
+@property (nonatomic, readonly) NSUInteger timestamp;
+
+/**
+ User permissions
+ */
+@property (nonatomic, copy, readonly) NSDictionary<NSString *, QNPermission *> *permissions;
+
+/**
+ All products
+ */
+@property (nonatomic, copy, readonly) NSDictionary<NSString *, QNProduct *> *products;
+
+/**
+ User products
+ */
+@property (nonatomic, copy, readonly) NSDictionary<NSString *, QNProduct *> *userPoducts;
+
+
+@end
+
+typedef void (^QNLaunchCompletionHandler)(QNLaunchResult *result, NSError  *_Nullable error) NS_SWIFT_NAME(Qonversion.LaunchCompletionHandler);
+
+typedef void (^QNPermissionCompletionHandler)(NSDictionary<NSString *, QNPermission*> *result, NSError  *_Nullable error) NS_SWIFT_NAME(Qonversion.PermissionCompletionHandler);
+
+typedef void (^QNPurchaseCompletionHandler)(NSDictionary<NSString *, QNPermission*> *result, NSError  *_Nullable error, BOOL cancelled) NS_SWIFT_NAME(Qonversion.PurchaseCompletionHandler);
+typedef void (^QNRestoreCompletionHandler)(NSDictionary<NSString *, QNPermission*> *result, NSError  *_Nullable error) NS_SWIFT_NAME(Qonversion.RestoreCompletionHandler);
+
+typedef void (^QNProductsCompletionHandler)(NSDictionary<NSString *, QNProduct *> *) NS_SWIFT_NAME(Qonversion.ProductsCompletionHandler);
