+++ Sources/Qonversion/QNLaunchResult.m
