+++ Sources/Qonversion/QNUtils.h
