+++ Sources/Qonversion/QNUtils.m
