+++ Sources/Qonversion/Qonversion.h
@@ -1,38 +1,10 @@
-#import <Foundation/Foundation.h>
-#import <QNProduct.h>
-#import <QNPermission.h>
-#import <QNLaunchResult.h>
+#import <StoreKit/StoreKit.h>
+
+#import "QNLaunchResult.h"
+#import "QNProduct.h"
+#import "QNPermission.h"
 
 NS_ASSUME_NONNULL_BEGIN
-
-typedef NS_ENUM(NSInteger, QNAttributionProvider) {
-  QNAttributionProviderAppsFlyer = 0,
-  QNAttributionProviderBranch,
-  QNAttributionProviderAdjust,
-  QNAttributionProviderApple
-} NS_SWIFT_NAME(Qonversion.AttributionProvider);
-
-typedef void (^QNLaunchCompletionHandler)(QNLaunchResult *result, NSError  *_Nullable error) NS_SWIFT_NAME(Qonversion.LaunchCompletionHandler);
-
-typedef void (^QNPermissionCompletionHandler)(NSDictionary<NSString *, QNPermission*> *result, NSError  *_Nullable error) NS_SWIFT_NAME(Qonversion.PermissionCompletionHandler);
-
-typedef void (^QNPurchaseCompletionHandler)(NSDictionary<NSString *, QNPermission*> *result, NSError  *_Nullable error, BOOL cancelled) NS_SWIFT_NAME(Qonversion.PurchaseCompletionHandler);
-typedef void (^QNRestoreCompletionHandler)(NSDictionary<NSString *, QNPermission*> *result, NSError  *_Nullable error) NS_SWIFT_NAME(Qonversion.RestoreCompletionHandler);
-
-typedef void (^QNProductsCompletionHandler)(NSDictionary<NSString *, QNProduct *> *) NS_SWIFT_NAME(Qonversion.ProductsCompletionHandler);
-
-/**
- Qonversion Defined User Properties
- We defined some common case properties and provided API for adding them
- @see [Product Center](https://qonversion.io/docs/defined-user-properties)
- */
-typedef NS_ENUM(NSInteger, QNProperty) {
-  QNPropertyEmail = 0,
-  QNPropertyName,
-  QNPropertyAppsFlyerUserID,
-  QNPropertyAdjustUserID,
-  QNPropertyKochavaDeviceID
-} NS_SWIFT_NAME(Qonversion.Property);
 
 @interface Qonversion : NSObject
